requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.23.0
pymongo>=4.5.0
bcrypt>=4.1.2
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
python-multipart>=0.0.9
vosk==0.3.45
SoundFile==0.12.1
numpy==2.1.1
requests>=2.31.0
tokenizers>=0.13.3
torch>=2.0.0
pydub>=0.25.1
reportlab>=4.0.0
scikit-learn>=1.3.0
transformers>=4.30.0

